refactor: split view.ts modules for search and visual lines
